// src/lib/database.types.ts
export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "14.5"
  }
  public: {
    Tables: {
// ...
      posts: {
        Row: {
          comments: number | null
          created_at: string | null
          date_posted: string | null
          id: string
          likes: number | null
          notes: string | null
          shares: number | null
          title: string | null
          updated_at: string | null
          user_id: string
          views: number | null
        }
        Insert: {
          comments?: number | null
          created_at?: string | null
          date_posted?: string | null
          id?: string
          likes?: number | null
          notes?: string | null
          shares?: number | null
          title?: string | null
          updated_at?: string | null
          user_id?: string
          views?: number | null
        }
        Update: {
          comments?: number | null
          created_at?: string | null
          date_posted?: string | null
          id?: string
          likes?: number | null
          notes?: string | null
          shares?: number | null
          title?: string | null
          updated_at?: string | null
          user_id?: string
          views?: number | null
        }
        Relationships: []
      }
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}